Move group users to separate component
